update version + moveout mig detection
