+++ python/ray/_private/ray_constants.py
@@ -480,5 +480,3 @@ RAY_TPU_MAX_CONCURRENT_CONNECTIONS_ENV_VAR = "RAY_TPU_MAX_CONCURRENT_ACTIVE_CONN
 
 
 RAY_NODE_IP_FILENAME = "node_ip_address.json"
-
-RAY_ENABLE_MIG_DETECTION_ENV_VAR = "RAY_ENABLE_MIG_DETECTION"
